Fix function name conflict.
Instance-level names conflict with global names in Swift Playgrounds. Clear up the conflict by using the past tense in the global function that returns a completed traversal and the present tense in the instance-level method that iterates the traversal result and "visits" each node head by applying to it the closure supplied.